BACKEND/requirements.txt
djangorestframework
django
django-cors-headers
djangorestframework-simplejwt